requirements.txt
APScheduler==3.0.1
Babel==1.3
Django==1.6.11
Graphillion==0.98
Jinja2==2.7.3
Markdown==2.6.1
MarkupSafe==0.23
MySQL-python==1.2.5
PyDNET==0.92
PyYAML==3.11
Pygments==2.0.2
South==1.0.2
Sphinx==1.3.1
Twisted==15.0.0
Whoosh==2.6.0
alabaster==0.7.3
amqp==1.4.6
anyjson==0.3.3
argparse==1.3.0
beautifulsoup4==4.3.2
billiard==3.3.0.19
bleach==1.4.1
bottle==0.12.8
celery==3.1.17
chardet==2.3.0
decorator==3.4.0
django-admin-bootstrapped==2.3.1
django-allauth==0.18.0
django-appconf==1.0.1
django-braces==1.4.0
django-celery==3.1.16
django-celery-email==1.1.1
django-compressor==1.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-filter==0.9.1
django-haystack==2.3.1
django-reversion==1.8.5
django-suit==0.2.12
djangorestframework==3.0.2
djoser==0.1.0
dnet==1.12
docutils==0.12
feedparser==5.1.3
futures==2.2.0
html5lib==0.999
httpie==0.8.0
httplib2==0.8
kombu==3.0.24
lxml==3.4.2
networkx==1.9.1
oauthlib==0.7.2
ordereddict==1.1
pbr==0.10.8
pcapy==0.10.6
pyparsing==2.0.3
pypcap==1.1.1
python-handler-socket==0.2.4
python-openid==2.2.5
pytz==2015.2
pyzmail==1.0.3
requests==2.6.0
requests-oauthlib==0.4.2
scapy==2.3.1
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.7
sqlparse==0.1.14
stevedore==1.3.0
tzlocal==1.1.2
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
wsgiref==0.1.2
xlrd==0.9.3
xmltodict==0.9.0
yolk==0.4.3
zope.interface==4.1.2
